Fix: Cast long to int.

+++ Java/src/net/xngo/utils/java/math/Hash.java
@@ -44,7 +44,7 @@ public class Hash
     // Optimization: Reduce the number of reads.
     //  8KB < bufferSize < 5MB
     // TODO: Use multiple of 4KB or 8KB
-    int bufferSize = file.length()/100; 
+    int bufferSize = (int)file.length()/100; 
     if(bufferSize>5242880) // 5242880 bytes = 5MB
       bufferSize=5242880;
     else if(bufferSize<8192) // 8192 bytes = 8 KB
